remove "ssl-key-log" feature
