+++ Cargo.toml
@@ -11,46 +11,25 @@ readme = "Readme.md"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[features]
-default = ["native-tls"]
-native-tls = [
-    "tungstenite/native-tls",
-    "async-tungstenite/async-native-tls",
-    "dep:native-tls",
-    "dep:async-native-tls",
-]
-ssl-key-log = [
-    "tungstenite/rustls-tls-native-roots",
-    "dep:rustls",
-    "dep:rustls-native-certs",
-    "async-tungstenite/async-tls",
-    "dep:old_rustls",
-    "dep:async-tls",
-]
-
 
 [dependencies]
-thiserror = "1.0.57"
+async-io = "2.3.1"
+async-lock = "3.3.0"
+async-native-tls = "0.5.0"
+async-std = "1.12.0"
+async-tungstenite = { version = "0.25.0", features = ["async-native-tls"] }
+base64 = "0.21.7"
+chrono = "0.4.34"
+futures-util = "0.3.30"
+http = "1.0.0"
+httparse = "1.8.0"
 isahc = { version = "1.7.2", features = ["json"] }
+native-tls = "0.2.11"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.114"
-chrono = "0.4.34"
+thiserror = "1.0.57"
+tungstenite = { version = "0.21.0", features = ["native-tls"] }
 uuid = { version = "1.7.0", features = ["fast-rng", "v4"] }
-tungstenite = "0.21.0"
-async-tungstenite = "0.25.0"
-futures-util = "0.3.30"
-async-lock = "3.3.0"
-async-io = "2.3.1"
-native-tls = { version = "0.2.11", optional = true }
-async-native-tls = { version = "0.5.0", optional = true }
-rustls = { version = "0.22.2", optional = true }
-rustls-native-certs = { version = "0.7.0", optional = true }
-old_rustls = { version = "0.21.10", optional = true, package = "rustls" }
-async-tls = { version = "0.13.0", optional = true }
-http = "1.0.0"
-base64 = "0.21.7"
-async-std = "1.12.0"
-httparse = "1.8.0"
 
 [dev-dependencies]
 smol = "2.0.0"
