+++ examples/synthesize_proxy_async.rs
@@ -22,13 +22,13 @@ fn main() {
                 synthesize(tts, &config).await;
 
                 let tts =
-                    connect_proxy_async("socks4://localhost:10808".parse().unwrap(), None, None)
+                    connect_proxy_async("socks4a://localhost:10808".parse().unwrap(), None, None)
                         .await
                         .unwrap();
                 synthesize(tts, &config).await;
 
                 let tts =
-                    connect_proxy_async("socks5://localhost:10808".parse().unwrap(), None, None)
+                    connect_proxy_async("socks5h://localhost:10808".parse().unwrap(), None, None)
                         .await
                         .unwrap();
                 synthesize(tts, &config).await;
