+++ examples/synthesize_stream_proxy.rs
@@ -28,12 +28,12 @@ fn main() {
             synthesize(sender, reader, config.clone());
 
             let (sender, reader) =
-                msedge_tts_split_proxy("socks4a://127.0.0.1:10808".parse().unwrap(), None, None)
+                msedge_tts_split_proxy("socks4://127.0.0.1:10808".parse().unwrap(), None, None)
                     .unwrap();
             synthesize(sender, reader, config.clone());
 
             let (sender, reader) =
-                msedge_tts_split_proxy("socks5h://127.0.0.1:10808".parse().unwrap(), None, None)
+                msedge_tts_split_proxy("socks5://127.0.0.1:10808".parse().unwrap(), None, None)
                     .unwrap();
             synthesize(sender, reader, config.clone());
             break;
