+++ src/tts/mod.rs
@@ -1,17 +1,14 @@
 //! TTS Client and Stream, SpeechConfig, Response Type.
-mod proxy;
 
 pub mod client;
 pub mod stream;
 
-#[cfg(not(feature = "ssl-key-log"))]
-#[path = "native_tls.rs"]
-mod tls;
-#[cfg(feature = "ssl-key-log")]
-#[path = "rustls_tls.rs"]
-mod tls;
-
-use crate::error::{Error, Result};
+mod proxy;
+use crate::error::{Error, ProxyError, Result};
+use proxy::{
+    http_proxy, http_proxy_async, socks4_proxy, socks4_proxy_async, socks5_proxy,
+    socks5_proxy_asnyc, ProxyAsyncStream, ProxyStream,
+};
 
 /// Synthesis Config
 #[derive(Debug)]
@@ -243,4 +240,87 @@ fn build_ssml_message(text: &str, config: &SpeechConfig) -> tungstenite::Message
         ssml,
     );
     tungstenite::Message::Text(ssml_message)
+}
+
+type WebSocketStream<T> = tungstenite::WebSocket<tungstenite::stream::MaybeTlsStream<T>>;
+
+fn websocket_connect() -> Result<WebSocketStream<std::net::TcpStream>> {
+    let request = build_websocket_request()?;
+    let (websocket, _) = tungstenite::connect(request)?;
+    Ok(websocket)
+}
+
+fn websocket_connect_proxy(
+    proxy: http::Uri,
+    username: Option<&str>,
+    password: Option<&str>,
+) -> Result<WebSocketStream<ProxyStream>> {
+    use tungstenite::handshake::HandshakeError;
+
+    let request = build_websocket_request()?;
+    let stream: std::result::Result<ProxyStream, ProxyError> = match proxy.scheme_str() {
+        Some(scheme) => match scheme.to_lowercase().as_str() {
+            "socks4" | "socks4a" => {
+                socks4_proxy(request.uri().host().unwrap(), proxy, username).map_err(|e| e.into())
+            }
+            "socks5" | "socks5h" => {
+                socks5_proxy(request.uri().host().unwrap(), proxy, username, password)
+                    .map_err(|e| e.into())
+            }
+            "http" | "https" => {
+                http_proxy(request.uri().host().unwrap(), proxy, username, password)
+                    .map_err(|e| e.into())
+            }
+            _ => Err(ProxyError::NotSupportedScheme(proxy)),
+        },
+        None => http_proxy(request.uri().host().unwrap(), proxy, username, password)
+            .map_err(|e| e.into()),
+    };
+    let (websocket, _) = tungstenite::client_tls(request, stream?).map_err(|e| match e {
+        HandshakeError::Failure(e) => e,
+        HandshakeError::Interrupted(_) => panic!("Bug: blocking handshake not blocked"),
+    })?;
+    Ok(websocket)
+}
+
+type WebSocketStreamAsync<T> =
+    async_tungstenite::WebSocketStream<async_tungstenite::async_std::ClientStream<T>>;
+
+async fn websocket_connect_async() -> Result<WebSocketStreamAsync<async_std::net::TcpStream>> {
+    let request = build_websocket_request()?;
+    let (websocket, _) = async_tungstenite::async_std::connect_async(request).await?;
+    Ok(websocket)
+}
+
+async fn websocket_connect_proxy_async(
+    proxy: http::Uri,
+    username: Option<&str>,
+    password: Option<&str>,
+) -> Result<WebSocketStreamAsync<ProxyAsyncStream>> {
+    let request = build_websocket_request()?;
+    let stream: std::result::Result<ProxyAsyncStream, ProxyError> = match proxy.scheme_str() {
+        Some(scheme) => match scheme.to_lowercase().as_str() {
+            "socks4" | "socks4a" => {
+                socks4_proxy_async(request.uri().host().unwrap(), proxy, username)
+                    .await
+                    .map_err(|e| e.into())
+            }
+            "socks5" | "socks5h" => {
+                socks5_proxy_asnyc(request.uri().host().unwrap(), proxy, username, password)
+                    .await
+                    .map_err(|e| e.into())
+            }
+            "http" | "https" => {
+                http_proxy_async(request.uri().host().unwrap(), proxy, username, password)
+                    .await
+                    .map_err(|e| e.into())
+            }
+            _ => Err(ProxyError::NotSupportedScheme(proxy)),
+        },
+        None => http_proxy_async(request.uri().host().unwrap(), proxy, username, password)
+            .await
+            .map_err(|e| e.into()),
+    };
+    let (websocket, _) = async_tungstenite::async_std::client_async_tls(request, stream?).await?;
+    Ok(websocket)
 }
